De-abstract the tokenscript functions out of the tokenscript library and into a TokenscriptFunctions class which is identical for both DMZ and App
