+++ app/build.gradle
@@ -38,8 +38,8 @@ android {
         applicationId "io.stormbird.wallet"
         minSdkVersion 18
         targetSdkVersion 28
-        versionCode 56
-        versionName "2.00.1"
+        versionCode 57
+        versionName "2.00.2"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         multiDexEnabled = true
 
@@ -165,8 +165,6 @@ dependencies {
     // PW
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation project(':tn')
-    
-    //implementation 'org.bouncycastle:bcprov-jdk15on:1.61'
 }
 
 // WARNING WARNING WARNING
