+++ app/src/main/java/io/stormbird/wallet/repository/TokenRepositoryType.java
@@ -37,6 +37,4 @@ public interface TokenRepositoryType {
 
     Disposable updateBlockRead(Token token, Wallet wallet);
     Single<String> resolveProxyAddress(TokenInfo tokenInfo);
-
-    String generateTransactionPayload(Token token, BigInteger tokenId, FunctionDefinition fd, TokenDefinition td);
 }
