+++ app/src/test/java/io/stormbird/wallet/MarketOrderTest.java
@@ -153,7 +153,7 @@ public class MarketOrderTest
                     try {
                         System.arraycopy(sigData.getR(), 0, sig, 0, 32);
                         System.arraycopy(sigData.getS(), 0, sig, 32, 32);
-                        sig[64] = sigData.getV()[0];
+                        sig[64] = sigData.getV();//[0];
                     }
                     catch (IndexOutOfBoundsException e)
                     {
