+++ app/src/test/java/io/stormbird/wallet/QRSelectionTest.java
@@ -107,7 +107,7 @@ public class QRSelectionTest
                     {
                         System.arraycopy(sigData.getR(), 0, sig, 0, 32);
                         System.arraycopy(sigData.getS(), 0, sig, 32, 32);
-                        sig[64] = sigData.getV()[0];
+                        sig[64] = sigData.getV();//[0];
                     }
                     catch (IndexOutOfBoundsException e)
                     {
