+++ app/src/test/java/io/stormbird/wallet/UniversalLinkTest.java
@@ -183,7 +183,7 @@ public class UniversalLinkTest
         try {
             System.arraycopy(sigData.getR(), 0, sig, 0, 32);
             System.arraycopy(sigData.getS(), 0, sig, 32, 32);
-            sig[64] = sigData.getV()[0];
+            sig[64] = sigData.getV();//[0];
         } catch (IndexOutOfBoundsException e) {
             throw new SalesOrderMalformed("Signature shorter than expected 256");
         }
