+++ lib/build.gradle
@@ -3,7 +3,6 @@ apply plugin: 'java-library'
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     testImplementation 'junit:junit:4.12'
-    implementation('org.web3j:core:4.3.0')
 }
 
 sourceCompatibility = "1.8"
